python/requirements.txt
aniso8601==1.1.0
click==6.6
Flask==0.11.1
Flask-HTTPAuth==3.1.2
Flask-RESTful==0.3.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
modbus-tk==0.5.0
pyserial==2.7
python-dateutil==2.5.3
pytz==2016.6.1
pyusb==1.0.0
requests==2.10.0
six==1.10.0
sortedcontainers==1.5.3
Werkzeug==0.11.10
xmltodict==0.10.2